Shift logic to angular services
