+++ src/main/angular/app/graph-processor.service.spec.ts
@@ -1,0 +1,16 @@
+/* import { TestBed } from '@angular/core/testing';
+
+import { GraphProcessorService } from './graph-processor.service';
+
+describe('GraphProcessorService', () => {
+  let service: GraphProcessorService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(GraphProcessorService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+}); */
