+++ src/main/angular/app/graph/graph.component.html
@@ -16,6 +16,6 @@ limitations under the License.
 
 <h1 class="center" id="project-prompt" *ngIf="!shouldShowChart">Please select a project below to generate a graph</h1>
 
-<google-chart #chart id="chart" class="center" *ngIf="shouldShowChart" [type]="type" [columns]="properties.columns"
-  [data]="properties.graphData" [options]="properties.options" [title]="title">
+<google-chart #chart id="chart" class="center" *ngIf="shouldShowChart" [type]="properties.type" [columns]="properties.columns"
+  [data]="properties.graphData" [options]="properties.options" [title]="properties.title">
 </google-chart>
